Kivy: update mesa to mesa_glu
This still doesn't work, because on the current master cython is on
version 0.28.2; and Kivy 1.10.0 is known-broken on Cython >= 0.27

See e.g. https://github.com/kivy/kivy/issues/5432

Once kivy hits 1.10.1, we might be able to move it forward. In the present
version of the installation notes (of Kivy) the following note might be
interpreted to indicate it will work:

Kivy 1.10.1.dev0 requires at least Cython version 0.23, and has been
tested through 0.27.3. Later versions may work, but as they have not
been tested there is no guarantee. The following versions of Cython have
known issues and cannot be used with Kivy: 0.27, 0.27.2